Resolve 9-file conflict set so the Windows build fix (PR #1592) can merge:
- CMakeLists_windows.txt: PR's superset source list (incl. main's new files)
- zinc_cpp CMake: keep main's if(MSVC) block (equivalent to PR's genex)
- zinc main.cpp: PR's manual argv scan (main's getopt_long still breaks MSVC)
- build_windows.cmd: main's vswhere -requires VCTools + PR's -DREPO_ROOT=%REPO_DIR%
  (main's %REPO_ROOT% is undefined in the script -> empty define)
- backend_onnx.cpp: PR's io.h/_access shim below covers both platforms
- q4nx_reader.cpp: PR's <cinttypes>
- gguf_to_onebp.cpp: PR's byte-order guard (drops main's broken static_assert
  referencing possibly-undefined __BYTE_ORDER__), main's PRIu64 formats,
  merged: PRIu64 + PR's if(inf) null guard
- unified_server.cpp: PR's std::filesystem scan + setvbuf _WIN32 guard
  (kept ONE_BIN_DISPATCH comment)

Syntax-checked with g++ -fsyntax-only and mingw cross on all touched files.

+++ engine/gpu/zinc_cpp/src/compute_engine.cpp
@@ -388,7 +388,7 @@ bool InferenceEngine::init(ComputeEngine& ce, ModelGPU& m) {
     alloc(hidden,     d.hidden * sizeof(float),                              "hidden");
     alloc(residual,   d.hidden * sizeof(float),                              "residual");
     alloc(qkv,        (d.n_heads * d.head_dim + d.n_kv_heads * d.head_dim * 2) * sizeof(float), "qkv");
-    alloc(attn_out,   d.n_heads * d.head_dim * sizeof(float),                 "attn_out");
+    alloc(attn_out,   (size_t)d.n_heads * d.head_dim * sizeof(float),                 "attn_out");
     alloc(gate_up,    d.inter * 2 * sizeof(float),                            "gate_up");
     alloc(silu_buf,   d.inter * sizeof(float),                                "silu_buf");
     alloc(logits,     d.vocab * sizeof(float),                                "logits");
@@ -497,4 +497,3 @@ int InferenceEngine::generate(int token_id) {
     pos++;
     return compute->argmax(logits.buffer(), d.vocab);
 }
-
